requirements.txt
# PhishGuard Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0
aiohttp==3.9.1
httpx==0.26.0
urllib3==2.1.0

# ML/Data Processing
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4
joblib==1.3.2

# HTML Parsing
beautifulsoup4==4.12.3
lxml==5.1.0

# String Matching
python-Levenshtein==0.23.0

# Scheduling
apscheduler==3.10.4

# DNS/WHOIS
dnspython==2.4.2
python-whois==0.9.4

# Security/Utilities
tld==0.13
